fixed issue id being long and int mixed up, now only using long. userservice has to save before other objects can access its id so needed to swap order in userservice

+++ server/src/main/java/com/codecool/sv_server/service/EmailService.java
@@ -26,7 +26,7 @@ public class EmailService {
         message.setFrom(email);
         mailSender.send(message);
     }
-    public void sendActivationTokenEmail(String token, String email, int id) {
+    public void sendActivationTokenEmail(String token, String email, long id) {
 
         String activationLink = "http://localhost:8080/api/auth/activate?token=" +
                                 token + "&userId=" + id;
